freemarker遍历key值为实体的map有点问题，只好在实体类中加了个属性

+++ sports/src/main/java/cn/sinoutx/sports/model/dao/EquipmentDao.java
@@ -12,7 +12,7 @@ import cn.sinoutx.sports.model.entity.Equipment;
 public interface EquipmentDao extends CrudRepository<Equipment, Integer> {
 
 	@Query("select count(*) from Equipment e where e.eqnid = ?1")
-	Integer countbyeqname(Eqname eqname);
+	Long countbyeqname(Eqname eqname);
 	
 	@Query("select e from Equipment e join e.eqnid n where n.eqtid = ?1")
 	List<Equipment> findByEqtype(Eqtype eqtid);
